@@ -11,7 +11,7 @@
     <GeneratePackageOnBuild Condition="'$(Configuration)' == 'Release'">True</GeneratePackageOnBuild>
     <ImplicitUsings>enable</ImplicitUsings>
     <IncludeSymbols>True</IncludeSymbols>
-    <LangVersion>10.0</LangVersion>
+    <LangVersion>12.0</LangVersion>
     <NoWarn>1591</NoWarn>
     <NeutralLanguage>en-us</NeutralLanguage>
     <Nullable>enable</Nullable>
@@ -20,13 +20,14 @@
     <RepositoryType>git</RepositoryType>
     <SignAssembly>True</SignAssembly>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PolySharpExcludeGeneratedTypes>System.Runtime.CompilerServices.IsExternalInit</PolySharpExcludeGeneratedTypes>
+    <PublicKey>0024000004800000940000000602000000240000525341310004000001000100b5808d2661c4f48da3683c5a557422a4ca5a132ac461790b148b8eabb94f109dad5ac0c3889f8e68f4923afb429cd5583ed0290540b5ec89a9153c954ad38f2e123ac292ca0f1b495a4bc103e5f7dec05264943f3ee018cb5e8feceff3659470a6837dc1f6c324e3b537245947032eba5c570e47a6573c5d75d9e2b12981f6b1</PublicKey>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ConfigureAwaitChecker.Analyzer" Version="5.*" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.*" PrivateAssets="All" />
+    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.*" PrivateAssets="All" Condition="'$(TargetFrameworkIdentifier)' == '.NETFramework'" />
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.*" PrivateAssets="All" />
-    <PackageReference Include="Nullable" Version="1.*" PrivateAssets="All" />
+    <PackageReference Include="PolySharp" Version="1.*" PrivateAssets="All" />
 
     <Reference Include="System.Net.Http" Condition="'$(TargetFrameworkIdentifier)' == '.NETFramework'" />
   </ItemGroup>
