@@ -2,12 +2,13 @@
 
   <PropertyGroup>
     <IsPackable>false</IsPackable>
+    <ParallelizationMode>Enhance</ParallelizationMode>
     <GenerateProgramFile>false</GenerateProgramFile>
-    <LangVersion>6.0</LangVersion>
+    <LangVersion>8.0</LangVersion>
     <DisableImplicitSystemValueTupleReference>true</DisableImplicitSystemValueTupleReference>
 
     <RollForward>Major</RollForward>
-    <TargetFrameworks>net6.0</TargetFrameworks>
+    <TargetFrameworks>net6.0;net8.0</TargetFrameworks>
     <TargetFrameworks Condition="'$(OS)' == 'Windows_NT'">net472;$(TargetFrameworks)</TargetFrameworks>
   </PropertyGroup>
 
