@@ -54,12 +54,13 @@ public static class Rules
         "Startup method parameter", DiagnosticSeverity.Error, true,
         CreateLocalizableString(nameof(AnalyzerResources.ConfigureServicesDescription)));
 
-    private static LocalizableResourceString CreateLocalizableString(string key) => new LocalizableResourceString(key, AnalyzerResources.ResourceManager, typeof(AnalyzerResources));
+    private static LocalizableResourceString CreateLocalizableString(string key) =>
+        new(key, AnalyzerResources.ResourceManager, typeof(AnalyzerResources));
 
     public static ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics { get; } = ImmutableArray.CreateRange(typeof(Rules)
         .GetProperties(BindingFlags.Public | BindingFlags.Static)
         .Where(p => p.PropertyType == typeof(DiagnosticDescriptor))
         .Select(p => p.GetMethod)
         .Where(m => m != null)
-        .Select(m => (DiagnosticDescriptor)m!.Invoke(null, Array.Empty<object?>())!));
+        .Select(m => (DiagnosticDescriptor)m!.Invoke(null, [])!));
 }
