@@ -21,13 +21,14 @@ public class SkippableFactTestCaseRunnerWrapper : DependencyInjectionTestCaseRun
     public override Type TestCaseType => typeof(SkippableFactTestCase);
 
     /// <inheritdoc />
-    public override async Task<RunSummary> RunAsync(IXunitTestCase testCase, IServiceProvider provider, IMessageSink diagnosticMessageSink, IMessageBus messageBus, object?[] constructorArguments, ExceptionAggregator aggregator, CancellationTokenSource cancellationTokenSource)
+    public override async Task<RunSummary> RunAsync(IXunitTestCase testCase, DependencyInjectionContext context,
+        IMessageSink diagnosticMessageSink, IMessageBus messageBus, object?[] constructorArguments,
+        ExceptionAggregator aggregator, CancellationTokenSource cancellationTokenSource)
     {
         var messageBusInterceptor = new SkippableTestMessageBus(messageBus, GetSkippingExceptionNames(testCase));
 
-        var result = await base.RunAsync(testCase, provider, diagnosticMessageSink,
-                messageBusInterceptor, constructorArguments, aggregator, cancellationTokenSource)
-            .ConfigureAwait(false);
+        var result = await base.RunAsync(testCase, context, diagnosticMessageSink,
+                messageBusInterceptor, constructorArguments, aggregator, cancellationTokenSource);
 
         result.Failed -= messageBusInterceptor.SkippedCount;
         result.Skipped += messageBusInterceptor.SkippedCount;
