@@ -10,29 +10,34 @@ public class UITestCaseRunnerAdapter : IXunitTestCaseRunnerWrapper
     /// <inheritdoc />
     public virtual Type TestCaseType => typeof(UITestCase);
 
-    public async Task<RunSummary> RunAsync(IXunitTestCase testCase, IServiceProvider provider, IMessageSink diagnosticMessageSink, IMessageBus messageBus, object?[] constructorArguments, ExceptionAggregator aggregator, CancellationTokenSource cancellationTokenSource)
+    public async Task<RunSummary> RunAsync(IXunitTestCase testCase, DependencyInjectionContext context,
+        IMessageSink diagnosticMessageSink, IMessageBus messageBus, object?[] constructorArguments,
+        ExceptionAggregator aggregator, CancellationTokenSource cancellationTokenSource)
     {
-        var scope = provider.CreateScope();
+        var scope = context.RootServices.CreateAsyncScope();
 
-        await using var _ = scope.AsAsyncDisposable().ConfigureAwait(false);
+        await using var _ = scope;
 
         var raw = new Dictionary<int, object>();
         foreach (var kv in FromServicesAttribute.CreateFromServices(testCase.Method.ToRuntimeMethod()))
         {
             raw[kv.Key] = testCase.TestMethodArguments[kv.Key];
 
-            testCase.TestMethodArguments[kv.Key] = kv.Value == typeof(ITestOutputHelper)
+            testCase.TestMethodArguments[kv.Key] = kv.Value.ParameterType == typeof(ITestOutputHelper)
                 ? throw new NotSupportedException("Can't inject ITestOutputHelper via method arguments when use StaFact")
-                : provider.GetService(kv.Value);
+                : scope.ServiceProvider.GetService(kv.Value);
         }
 
-        constructorArguments = DependencyInjectionTestMethodRunner.CreateTestClassConstructorArguments(scope.ServiceProvider, constructorArguments, aggregator);
+        constructorArguments = scope.ServiceProvider.CreateTestClassConstructorArguments(constructorArguments, aggregator);
 
-        var summary = await testCase.RunAsync(diagnosticMessageSink, messageBus, constructorArguments, aggregator, cancellationTokenSource).ConfigureAwait(false);
-
-        foreach (var kv in raw)
-            testCase.TestMethodArguments[kv.Key] = kv.Value;
-
-        return summary;
+        try
+        {
+            return await testCase.RunAsync(diagnosticMessageSink, messageBus, constructorArguments, aggregator, cancellationTokenSource);
+        }
+        finally
+        {
+            foreach (var kv in raw)
+                testCase.TestMethodArguments[kv.Key] = kv.Value;
+        }
     }
 }
