@@ -10,17 +10,16 @@ public class RetryTestCaseRunnerWrapper : DependencyInjectionTestCaseRunnerWrapp
     public override Type TestCaseType => typeof(RetryTestCase);
 
     /// <inheritdoc />
-    public override Task<RunSummary> RunAsync(IXunitTestCase testCase, IServiceProvider provider, IMessageSink diagnosticMessageSink,
-        IMessageBus messageBus, object?[] constructorArguments, ExceptionAggregator aggregator,
-        CancellationTokenSource cancellationTokenSource)
+    public override Task<RunSummary> RunAsync(IXunitTestCase testCase, DependencyInjectionContext context,
+        IMessageSink diagnosticMessageSink, IMessageBus messageBus, object?[] constructorArguments,
+        ExceptionAggregator aggregator, CancellationTokenSource cancellationTokenSource)
     {
         if (testCase is not IRetryableTestCase retryableTestCase)
-        {
             throw new ArgumentException("Must be a retryable test case", nameof(testCase));
-        }
 
-        return RetryTestCaseRunner.RunAsync(retryableTestCase, diagnosticMessageSink, messageBus, cancellationTokenSource,
-            blockingMessageBus => base.RunAsync(retryableTestCase, provider, diagnosticMessageSink, blockingMessageBus,
+        return RetryTestCaseRunner.RunAsync(retryableTestCase, diagnosticMessageSink, messageBus,
+            cancellationTokenSource,
+            blockingMessageBus => base.RunAsync(retryableTestCase, context, diagnosticMessageSink, blockingMessageBus,
                 constructorArguments, aggregator, cancellationTokenSource));
     }
 }
