@@ -1,33 +1,23 @@
 ﻿namespace Xunit.DependencyInjection;
 
-public class DependencyInjectionTestCollectionRunner : XunitTestCollectionRunner
+public class DependencyInjectionTestCollectionRunner(
+    DependencyInjectionStartupContext context,
+    ITestCollection testCollection,
+    IEnumerable<IXunitTestCase> testCases,
+    IMessageSink diagnosticMessageSink,
+    IMessageBus messageBus,
+    ITestCaseOrderer testCaseOrderer,
+    ExceptionAggregator aggregator,
+    CancellationTokenSource cancellationTokenSource)
+    : XunitTestCollectionRunner(testCollection, testCases, diagnosticMessageSink,
+        messageBus, testCaseOrderer, aggregator, cancellationTokenSource)
 {
-    private readonly IServiceProvider? _provider;
-    private readonly IReadOnlyDictionary<ITestClass, IHost?> _hostMap;
-    private IServiceScope? _serviceScope;
-    private readonly IMessageSink _diagnosticMessageSink;
-
-    public DependencyInjectionTestCollectionRunner(IServiceProvider? provider,
-        ITestCollection testCollection,
-        IEnumerable<IXunitTestCase> testCases,
-        IReadOnlyDictionary<ITestClass, IHost?> hostMap,
-        IMessageSink diagnosticMessageSink,
-        IMessageBus messageBus,
-        ITestCaseOrderer testCaseOrderer,
-        ExceptionAggregator aggregator,
-        CancellationTokenSource cancellationTokenSource)
-        : base(testCollection, testCases, diagnosticMessageSink,
-            messageBus, testCaseOrderer, aggregator, cancellationTokenSource)
-    {
-        _provider = provider;
-        _hostMap = hostMap;
-        _diagnosticMessageSink = diagnosticMessageSink;
-    }
+    private AsyncServiceScope? _serviceScope;
 
     /// <inheritdoc />
     protected override void CreateCollectionFixture(Type fixtureType)
     {
-        if (_provider == null)
+        if (context.DefaultRootServices is not { } provider)
         {
             Aggregator.Add(HostManager.MissingDefaultHost("Collection fixture  require a default startup."));
 
@@ -35,32 +25,37 @@ public class DependencyInjectionTestCollectionRunner : XunitTestCollectionRunner
         }
         else
         {
-            _serviceScope = _provider.GetRequiredService<IServiceScopeFactory>().CreateScope();
+            var serviceScope = provider.GetRequiredService<IServiceScopeFactory>().CreateAsyncScope();
+
+            _serviceScope = serviceScope;
 
             Aggregator.Run(() => CollectionFixtureMappings[fixtureType] =
-                ActivatorUtilities.GetServiceOrCreateInstance(_serviceScope.ServiceProvider, fixtureType));
+                ActivatorUtilities.GetServiceOrCreateInstance(serviceScope.ServiceProvider, fixtureType));
         }
-
     }
 
     /// <inheritdoc/>
     protected override async Task BeforeTestCollectionFinishedAsync()
     {
-        await base.BeforeTestCollectionFinishedAsync().ConfigureAwait(false);
+        await base.BeforeTestCollectionFinishedAsync();
 
         foreach (var fixture in CollectionFixtureMappings.Values.OfType<IAsyncDisposable>())
-            await Aggregator.RunAsync(() => fixture.DisposeAsync().AsTask()).ConfigureAwait(false);
+            await Aggregator.RunAsync(() => fixture.DisposeAsync().AsTask());
 
-        if (_serviceScope != null) await _serviceScope.DisposeAsync().ConfigureAwait(false);
+        if (_serviceScope is { } disposable) await disposable.DisposeAsync();
     }
 
     /// <inheritdoc />
     protected override Task<RunSummary> RunTestClassAsync(ITestClass testClass,
         IReflectionTypeInfo @class, IEnumerable<IXunitTestCase> testCases) =>
-        _hostMap.TryGetValue(testClass, out var host) && host != null
-            ? new DependencyInjectionTestClassRunner(host.Services, testClass, @class, testCases,
-                    _diagnosticMessageSink, MessageBus, TestCaseOrderer,
-                    new(Aggregator), CancellationTokenSource, CollectionFixtureMappings)
-                .RunAsync()
+        context.ContextMap.TryGetValue(testClass, out var value) && value != null
+            ? new DependencyInjectionTestClassRunner(
+                new(value.Host, value.DisableParallelization ||
+                    !(context.ParallelizationMode == ParallelizationMode.Force ||
+                        context.ParallelizationMode == ParallelizationMode.Enhance &&
+                        SynchronizationContext.Current is MaxConcurrencySyncContext),
+                    context.ParallelizationMode == ParallelizationMode.Force), testClass, @class, testCases,
+                DiagnosticMessageSink, MessageBus, TestCaseOrderer, new(Aggregator), CancellationTokenSource,
+                CollectionFixtureMappings).RunAsync()
             : base.RunTestClassAsync(testClass, @class, testCases);
 }
