@@ -1,21 +1,19 @@
 ﻿namespace Xunit.DependencyInjection;
 
 [AttributeUsage(AttributeTargets.Parameter)]
-public class FromServicesAttribute : Attribute
+public class FromServicesAttribute() : FromKeyedServicesAttribute(null!)
 {
-    internal static IReadOnlyDictionary<int, Type> CreateFromServices(MethodInfo method)
+    internal static IReadOnlyDictionary<int, ParameterInfo> CreateFromServices(MethodInfo method)
     {
-        var dic = new Dictionary<int, Type>();
+        var dic = new Dictionary<int, ParameterInfo>();
 
         var parameters = method.GetParameters();
         for (var index = 0; index < parameters.Length; index++)
         {
-            var type = parameters[index].ParameterType;
-            if ((type.IsClass || type.IsInterface) &&
-                parameters[index].GetCustomAttribute<FromServicesAttribute>() != null)
-            {
-                dic[index] = type;
-            }
+            var parameter = parameters[index];
+            if (parameter.ParameterType is { IsClass: false, IsInterface: false }) continue;
+
+            if (parameter.GetCustomAttribute<FromKeyedServicesAttribute>() != null) dic[index] = parameter;
         }
 
         return dic;
