@@ -8,15 +8,14 @@ namespace Microsoft.Extensions.Internal;
 /// return, regardless of whether the underlying value is a System.Task, an FSharpAsync, or an
 /// application-defined custom awaitable.
 /// </summary>
-internal readonly struct ObjectMethodExecutorAwaitable
+internal readonly struct ObjectMethodExecutorAwaitable(
+    object customAwaitable,
+    Func<object, object> getAwaiterMethod,
+    Func<object, bool> isCompletedMethod,
+    Func<object, object> getResultMethod,
+    Action<object, Action> onCompletedMethod,
+    Action<object, Action> unsafeOnCompletedMethod)
 {
-    private readonly object _customAwaitable;
-    private readonly Func<object, object> _getAwaiterMethod;
-    private readonly Func<object, bool> _isCompletedMethod;
-    private readonly Func<object, object> _getResultMethod;
-    private readonly Action<object, Action> _onCompletedMethod;
-    private readonly Action<object, Action> _unsafeOnCompletedMethod;
-
     // Perf note: since we're requiring the customAwaitable to be supplied here as an object,
     // this will trigger a further allocation if it was a value type (i.e., to box it). We can't
     // fix this by making the customAwaitable type generic, because the calling code typically
@@ -35,52 +34,22 @@ internal readonly struct ObjectMethodExecutorAwaitable
     // We can reconsider this in the future if there's a need to optimize for ValueTask<T>
     // or other value-typed awaitables.
 
-    public ObjectMethodExecutorAwaitable(
-        object customAwaitable,
-        Func<object, object> getAwaiterMethod,
+    public Awaiter GetAwaiter() => new(getAwaiterMethod(customAwaitable), isCompletedMethod, getResultMethod,
+        onCompletedMethod, unsafeOnCompletedMethod);
+
+    public readonly struct Awaiter(
+        object customAwaiter,
         Func<object, bool> isCompletedMethod,
         Func<object, object> getResultMethod,
         Action<object, Action> onCompletedMethod,
         Action<object, Action> unsafeOnCompletedMethod)
+        : ICriticalNotifyCompletion
     {
-        _customAwaitable = customAwaitable;
-        _getAwaiterMethod = getAwaiterMethod;
-        _isCompletedMethod = isCompletedMethod;
-        _getResultMethod = getResultMethod;
-        _onCompletedMethod = onCompletedMethod;
-        _unsafeOnCompletedMethod = unsafeOnCompletedMethod;
-    }
+        public bool IsCompleted => isCompletedMethod(customAwaiter);
 
-    public Awaiter GetAwaiter() => new(_getAwaiterMethod(_customAwaitable), _isCompletedMethod, _getResultMethod,
-        _onCompletedMethod, _unsafeOnCompletedMethod);
+        public object GetResult() => getResultMethod(customAwaiter);
 
-    public readonly struct Awaiter : ICriticalNotifyCompletion
-    {
-        private readonly object _customAwaiter;
-        private readonly Func<object, bool> _isCompletedMethod;
-        private readonly Func<object, object> _getResultMethod;
-        private readonly Action<object, Action> _onCompletedMethod;
-        private readonly Action<object, Action> _unsafeOnCompletedMethod;
-
-        public Awaiter(
-            object customAwaiter,
-            Func<object, bool> isCompletedMethod,
-            Func<object, object> getResultMethod,
-            Action<object, Action> onCompletedMethod,
-            Action<object, Action> unsafeOnCompletedMethod)
-        {
-            _customAwaiter = customAwaiter;
-            _isCompletedMethod = isCompletedMethod;
-            _getResultMethod = getResultMethod;
-            _onCompletedMethod = onCompletedMethod;
-            _unsafeOnCompletedMethod = unsafeOnCompletedMethod;
-        }
-
-        public bool IsCompleted => _isCompletedMethod(_customAwaiter);
-
-        public object GetResult() => _getResultMethod(_customAwaiter);
-
-        public void OnCompleted(Action continuation) => _onCompletedMethod(_customAwaiter, continuation);
+        public void OnCompleted(Action continuation) => onCompletedMethod(customAwaiter, continuation);
 
         public void UnsafeOnCompleted(Action continuation)
         {
@@ -97,7 +66,7 @@ internal readonly struct ObjectMethodExecutorAwaitable
             //   if a caller sees that the proxy implements ICriticalNotifyCompletion but the proxy chooses to
             //   pass the call on to the underlying awaitable's OnCompleted method.
 
-            (_unsafeOnCompletedMethod ?? _onCompletedMethod)(_customAwaiter, continuation);
+            (unsafeOnCompletedMethod ?? onCompletedMethod)(customAwaiter, continuation);
         }
     }
 }
