@@ -1,11 +1,9 @@
 ﻿namespace Xunit.DependencyInjection;
 
 [AttributeUsage(AttributeTargets.Class)]
-public sealed class StartupAttribute : Attribute
+public sealed class StartupAttribute(Type startupType) : Attribute
 {
-    public StartupAttribute(Type startupType) => StartupType = startupType;
-
-    public Type StartupType { get; }
+    public Type StartupType { get; } = startupType;
 
     /// <summary>Default is true. If false, a isolated Startup will be created for the test class.</summary>
     public bool Shared { get; set; } = true;
