@@ -1,3 +1,3 @@
 ﻿namespace Xunit.DependencyInjection.Analyzer.Test;
 
-public class Startup { }
+public class Startup;
