@@ -1,20 +1,21 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Hosting.Server;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
+using Xunit;
+using Xunit.DependencyInjection.AspNetCoreTesting;
 using Xunit.DependencyInjection.Logging;
 
+[assembly: CollectionBehavior(DisableTestParallelization = true)]
 namespace Xunit.DependencyInjection.Test.AspNetCore;
 
 public class Startup
 {
     public void ConfigureHost(IHostBuilder hostBuilder) =>
         hostBuilder.ConfigureWebHost(webHostBuilder => webHostBuilder
-            .UseTestServer(options => options.PreserveExecutionContext = true)
+            .UseTestServerAndAddDefaultHttpClient()
             .UseStartup<AspNetCoreStartup>());
 
     private class AspNetCoreStartup
