@@ -1,37 +1,30 @@
 ﻿namespace Xunit.DependencyInjection.Test.ClassFixture;
 
 [TestCaseOrderer("Xunit.DependencyInjection.Test." + nameof(TestCaseByMethodNameOrderer), "Xunit.DependencyInjection.Test")]
-public class ClassFixtureTest : IClassFixture<FixtureWithDependency>
+public class ClassFixtureTest(FixtureWithDependency fixture) : IClassFixture<FixtureWithDependency>
 {
-    private readonly FixtureWithDependency _fixture;
-
-    public ClassFixtureTest(FixtureWithDependency fixture)
-    {
-        _fixture = fixture;
-    }
-
     [Fact]
     public void FixtureWithDependencyIsInjected()
     {
-        Assert.NotNull(_fixture);
+        Assert.NotNull(fixture);
     }
 
     [Fact]
     public void ClassFixtureContainsInjectedDependency()
     {
-        Assert.IsType<DependencyClass>(_fixture.Dependency);
+        Assert.IsType<DependencyClass>(fixture.Dependency);
     }
 
     [Fact]
     public void SameClassFixtureDependencyInstance_1()
     {
-        Assert.Equal(0, _fixture.Dependency.Value);
-        _fixture.Dependency.Value = 5555;
+        Assert.Equal(0, fixture.Dependency.Value);
+        fixture.Dependency.Value = 5555;
     }
 
     [Fact]
     public void SameClassFixtureDependencyInstance_2()
     {
-        Assert.Equal(5555, _fixture.Dependency.Value);
+        Assert.Equal(5555, fixture.Dependency.Value);
     }
 }
