@@ -1,12 +1,10 @@
 ﻿namespace Xunit.DependencyInjection.Test.CollectionFixture;
 
-public class CollectionFixtureWithDependency : IDisposable
+public class CollectionFixtureWithDependency(IDependency dependency) : IDisposable
 {
     public bool IsDisposed { get; private set; }
 
-    public CollectionFixtureWithDependency(IDependency dependency) => Dependency = dependency;
-
-    public IDependency Dependency { get; }
+    public IDependency Dependency { get; } = dependency;
 
     public void Dispose() => IsDisposed = true;
 }
