@@ -1,44 +1,43 @@
-﻿namespace Xunit.DependencyInjection.Test;
+﻿using Xunit.DependencyInjection.Demystifier;
 
-public class FromServicesTest
+namespace Xunit.DependencyInjection.Test;
+
+public class FromServicesTest(IDependency dependency, ITestOutputHelper output)
 {
-    private readonly IDependency _dependency;
-    private readonly ITestOutputHelper _output;
-
-    public FromServicesTest(IDependency dependency, ITestOutputHelper output)
-    {
-        _dependency = dependency;
-        _output = output;
-    }
-
     [Theory]
     [InlineData("Test", 3, null, null)]
     public void FactTest(string arg1,
         [FromServices]int invalid,
-        [FromServices]IDependency dependency,
-        [FromServices]ITestOutputHelper output)
+        [FromServices]IDependency dependency1,
+        [FromServices]ITestOutputHelper output1)
     {
         Assert.Equal("Test", arg1);
         Assert.Equal(3, invalid);
-        Assert.Equal(_dependency, dependency);
-        Assert.Equal(_output, output);
+        Assert.Equal(dependency, dependency1);
+        Assert.Equal(output, output1);
     }
 
     [Theory]
     [MemberData(nameof(TheoryData))]
     public void TheoryTest(string arg1,
-        [FromServices]int invalid,
-        [FromServices]IDependency dependency,
-        [FromServices]ITestOutputHelper output)
+        [FromServices] int invalid,
+        [FromServices] IDependency dependency1,
+        [FromServices] IAsyncExceptionFilter filter,
+        [FromKeyedServices("small")] IFromKeyedServicesTest test1,
+        [FromKeyedServices("large")] IFromKeyedServicesTest test2,
+        [FromServices] ITestOutputHelper output1)
     {
         Assert.Equal("Test", arg1);
         Assert.Equal(3, invalid);
-        Assert.Equal(_dependency, dependency);
-        Assert.Equal(_output, output);
+        Assert.Equal(dependency, dependency1);
+        Assert.IsType<DemystifyExceptionFilter>(filter);
+        Assert.IsType<FromSmallKeyedServicesTest>(test1);
+        Assert.IsType<FromLargeKeyedServicesTest>(test2);
+        Assert.Equal(output, output1);
     }
 
     public static IEnumerable<object?[]> TheoryData()
     {
-        yield return new object?[] { "Test", 3, null, null };
+        yield return ["Test", 3, null, null, null, null, null];
     }
 }
