@@ -4,8 +4,6 @@ public class ParameterlessConstructorTest
 {
     [Theory]
     [InlineData(null)]
-    public void ITestOutputHelperAccessor_Output_Should_Not_Null([FromServices]ITestOutputHelperAccessor accessor)
-    {
+    public void ITestOutputHelperAccessor_Output_Should_Not_Null([FromServices] ITestOutputHelperAccessor accessor) =>
         Assert.NotNull(accessor.Output);
-    }
 }
