@@ -1,4 +1,6 @@
-﻿namespace Xunit.DependencyInjection.Test;
+﻿using Xunit.DependencyInjection.Demystifier;
+
+namespace Xunit.DependencyInjection.Test;
 
 public class StaFactTest : IDisposable, IAsyncLifetime
 {
@@ -53,5 +55,16 @@ public class StaFactTest : IDisposable, IAsyncLifetime
         Assert.Equal(_ctorThreadId, Environment.CurrentManagedThreadId);
         Assert.Same(_ctorSyncContext, SynchronizationContext.Current);
         Assert.Equal(0, arg);
+    }
+
+    [UITheory]
+    [InlineData(null, null, null)]
+    public void FromKeyedServicesTest([FromServices] IAsyncExceptionFilter filter,
+        [FromKeyedServices("small")] IFromKeyedServicesTest test1,
+        [FromKeyedServices("large")] IFromKeyedServicesTest test2)
+    {
+        Assert.IsType<DemystifyExceptionFilter>(filter);
+        Assert.IsType<FromSmallKeyedServicesTest>(test1);
+        Assert.IsType<FromLargeKeyedServicesTest>(test2);
     }
 }
