@@ -3,6 +3,6 @@
 public class TestCaseByMethodNameOrderer : ITestCaseOrderer
 {
     public IEnumerable<TTestCase> OrderTestCases<TTestCase>(IEnumerable<TTestCase> testCases)
-        where TTestCase : ITestCase
-        => testCases.OrderBy(t => t.TestMethod.Method.Name);
+        where TTestCase : ITestCase =>
+        testCases.OrderBy(t => t.TestMethod.Method.Name);
 }
