@@ -2,9 +2,8 @@
 
 namespace Xunit.DependencyInjection.Test;
 
-public class XRetryTest
+public class XRetryTest(IDependency dependency)
 {
-    private readonly IDependency _dependency;
     private static int _factNumCalls = 0;
     // testId => numCalls
     private static readonly Dictionary<int, int> _theoryNumCalls = new()
@@ -18,11 +17,6 @@ public class XRetryTest
         { 0, 0 },
         { 1, 0 }
     };
-
-    public XRetryTest(IDependency dependency)
-    {
-        _dependency = dependency;
-    }
 
     [RetryFact]
     public void RetryFact()
@@ -44,14 +38,9 @@ public class XRetryTest
         Assert.Equal(3, _theoryNumCalls[id]);
     }
 
-    public class NonSerializableTestData
+    public class NonSerializableTestData(int id)
     {
-        public readonly int Id;
-
-        public NonSerializableTestData(int id)
-        {
-            Id = id;
-        }
+        public readonly int Id = id;
     }
 
     [RetryTheory]
@@ -64,18 +53,18 @@ public class XRetryTest
         Assert.Equal(3, _nonSerializableTheoryNumCalls[nonSerializableWrapper.Id]);
     }
 
-    public static IEnumerable<object[]> GetTestData() => new[]
-    {
-        new object[] { new NonSerializableTestData(0) },
-        new object[] { new NonSerializableTestData(1) }
-    };
+    public static IEnumerable<object[]> GetTestData() =>
+    [
+        [new NonSerializableTestData(0)],
+        [new NonSerializableTestData(1)]
+    ];
 
     private void AssertDependencyUnique()
     {
         // Assert dependency is in its original state
-        Assert.Equal(0, _dependency.Value);
+        Assert.Equal(0, dependency.Value);
 
         // Modify the dependency so that a retry of this test case would fail if it got the same instance
-        _dependency.Value++;
+        dependency.Value++;
     }
 }
