@@ -1,16 +1,12 @@
 ﻿namespace Xunit.DependencyInjection.Test;
 
-public class XunitTestOutputLoggerTest
+public class XunitTestOutputLoggerTest(ILogger<XunitTestOutputLoggerTest> logger)
 {
-    private readonly ILogger<XunitTestOutputLoggerTest> _logger;
-
-    public XunitTestOutputLoggerTest(ILogger<XunitTestOutputLoggerTest> logger) => _logger = logger;
-
     [Fact]
     public void Test()
     {
-        _logger.LogDebug("LogDebug");
-        _logger.LogInformation("LogInformation");
-        _logger.LogError("LogError");
+        logger.LogDebug("LogDebug");
+        logger.LogInformation("LogInformation");
+        logger.LogError("LogError");
     }
 }
